+++ denso_robot_moveit_config/CHANGELOG.rst
@@ -2,6 +2,10 @@
 Changelog for package denso_robot_moveit_config
 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
 
+3.3.0 (2023-02-14)
+------------------
+* Add parameter to specify bcap_slave_mode at startup
+
 3.2.0 (2021-06-02)
 ------------------